make a black list of observing run to skip
In the demo/tutorial ESO p2 Tool used for testing, there exists
some ObservingRun objects returned by the API, but not present in
the p2 GUI. (ESO appears to have filtered them). That makes sense
b/c they are missing components (obx containers) that make them
problematic to handle. The black list is list of these observing
run IDs, and we skip them in the ESO Facility. This a dev/testing
work around and shouldn't be needed for production data.